Revert document move and move login page to /login
